refactor(admin): centralize all-creators UI config
Moved duplicated all-creators configuration arrays into a shared admin utils module.
This keeps UI config in one place and leaves components focused on rendering.

Constraint: Keep existing rendering behavior and style components unchanged
Rejected: Split into two separate config files | unnecessary fragmentation
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Keep future table/detail field config updates in this shared utils module
Tested: Manual review of imports/exports and moved config integrity
Not-tested: Runtime UI interaction and full test suite
